Added user panel

+++ 4fitter/Controllers/BookmarksController.cs
@@ -3,6 +3,7 @@ using System.Web.Mvc;
 using _4fitter.Models;
 using Microsoft.AspNet.Identity;
 using Microsoft.AspNet.Identity.EntityFramework;
+using System.Linq;
 
 namespace _4fitter.Controllers
 {
@@ -14,8 +15,7 @@ namespace _4fitter.Controllers
         [HttpPost]
         public HttpStatusCode Create([Bind(Include = "ID,UserID,ArticleID")] Bookmark bookmark)
         {
-            var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
-            var userExists = userManager.FindById(bookmark.UserID) != null;
+            var userExists = this.IsUserExist(bookmark.UserID);
 
             if (ModelState.IsValid && userExists)
             {
@@ -27,17 +27,52 @@ namespace _4fitter.Controllers
             return HttpStatusCode.BadRequest;
         }
 
-        // POST: Bookmarks/Delete/5
-        [HttpPost, ActionName("Delete")]
-        [ValidateAntiForgeryToken]
-        public HttpStatusCode DeleteConfirmed(int id)
+        // GET: Bookmarks/Check/6
+        [HttpGet, ActionName("Check")]
+        public bool IsArticleBookmarked(int id, string userId)
         {
-            Bookmark bookmark = db.Bookmarks.Find(id);
+            return db.Bookmarks.Any(b => b.Article.ID == id && b.UserID == userId);
+        }
+
+        // POST: Bookmarks/Delete/5?userId=<guid>
+        [HttpPost, ActionName("Delete")]
+        public HttpStatusCode DeleteConfirmed(int id, string userId)
+        {
+            var userExists = this.IsUserExist(userId);
+
+            if (!userExists)
+            {
+                return HttpStatusCode.InternalServerError;
+            }
+
+            Bookmark bookmark = db.Bookmarks.First(b => b.Article.ID == id && b.UserID == userId);
             db.Bookmarks.Remove(bookmark);
 
             var retCode = db.SaveChanges();
 
             return retCode > 0 ? HttpStatusCode.OK : HttpStatusCode.InternalServerError;
+        }
+
+        // GET: Bookmarks/User/<guid>
+        [HttpGet, ActionName("User")]
+        public ActionResult GetBookmarksByUserId(string userId)
+        {
+            var userExists = this.IsUserExist(userId);
+
+            if (!userExists)
+            {
+                return new HttpStatusCodeResult(500);
+            }
+
+            var result = db.Bookmarks.Where(b => b.UserID == userId).ToList();
+
+            return Json(result, JsonRequestBehavior.AllowGet);
+        }
+
+        private bool IsUserExist(string id)
+        {
+            var userManager = new UserManager<ApplicationUser>(new UserStore<ApplicationUser>(db));
+            return userManager.FindById(id) != null;
         }
 
         protected override void Dispose(bool disposing)
